# doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Audio Files Hosting API Specs",
    "version": "1.0"
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "Get a list of users",
        "description": "Get a list of users",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "The page number to retrieve",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "description": "The number of users to retrieve per page",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "name": {
                            "type": "string",
                            "example": "John Doe"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2023-10-01T12:00:00Z"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2023-10-01T12:00:00Z"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "totalCount": {
                      "type": "integer",
                      "example": 100
                    }
                  },
                  "required": [
                    "items",
                    "totalCount"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch users"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (e.g., session expired).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Session expired"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (e.g., user is not an admin).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Forbidden to create new user"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "description": "Create a new user with a unique username and password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The username for the new user.",
                    "example": "johndoe"
                  },
                  "password": {
                    "type": "string",
                    "description": "The password for the new user.",
                    "example": "securepassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User created successfully"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "username": {
                          "type": "string",
                          "example": "johndoe"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2023-10-01T12:00:00Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2023-10-01T12:00:00Z"
                        }
                      },
                      "required": [
                        "id",
                        "username",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "message",
                    "user"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Failed to create user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "username": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "example": "Username already exists"
                          }
                        },
                        "password": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "example": "Password must be at least 3 characters long"
                          }
                        },
                        "server": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "example": "Failed to create user"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (e.g., session expired).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "server": {
                          "type": "string",
                          "example": "Session expired"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (e.g., user is not an admin).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "server": {
                          "type": "string",
                          "example": "Forbidden to create new user"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}